requirements.txt
Flask>=2.0.0
Flask-CORS>=3.0.0
requests>=2.28.0
PyJWT>=2.0.0
bcrypt>=4.0.0
python-dotenv>=1.0.0
Werkzeug>=2.0.0
numpy>=1.20.0